processinfo repackaged into newprocessinfo
