+++ newprocessinfo/pinfo_common_defs.h
@@ -1,0 +1,4 @@
+#ifndef _PINFO_COMMON_DEFS_H
+#define _PINFO_COMMON_DEFS_H
+
+#endif // #ifndef _PINFO_COMMON_DEFS_H
