+++ newprocessinfo/timeutils.h
@@ -44,4 +44,4 @@ double timespec_diff_double(struct timespec start, struct timespec end);
 
 char *timedouble_to_UTC_timeofdaystring(double timedouble);
 
-#endif
+#endif // #ifndef _CLICORE_TIMEUTILS_H
