--- processinfo/processinfo_SIGexit.h
@@ -1,6 +1,0 @@
-#ifndef _PROCESSINFO_SIGEXIT_H
-#define _PROCESSINFO_SIGEXIT_H
-
-int processinfo_SIGexit(PROCESSINFO *processinfo, int SignalNumber);
-
-#endif
