--- processinfo/processinfo_WriteMessage.h
@@ -1,6 +1,0 @@
-#ifndef _PROCESSINFO_WRITEMESSAGE_H
-#define _PROCESSINFO_WRITEMESSAGE_H
-
-int processinfo_WriteMessage(PROCESSINFO *processinfo, const char *msgstring);
-
-#endif
