--- processinfo/processinfo_exec_end.h
@@ -1,6 +1,0 @@
-#ifndef _PROCESSINFO_EXEC_END_H
-#define _PROCESSINFO_EXEC_END_H
-
-int processinfo_exec_end(PROCESSINFO *processinfo);
-
-#endif
