--- processinfo/processinfo_exec_start.h
@@ -1,6 +1,0 @@
-#ifndef _PROCESSINFO_EXEC_START_H
-#define _PROCESSINFO_EXEC_START_H
-
-int processinfo_exec_start(PROCESSINFO *processinfo);
-
-#endif
