--- processinfo/processinfo_internal.h
@@ -1,12 +1,0 @@
-#ifndef PROCESSINFO_INTERNAL_H
-#define PROCESSINFO_INTERNAL_H
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <errno.h>
-
-#include "ImageStreamIO/milkDebugTools.h"
-#include "ImageStreamIO/ImageStruct.h"
-
-#endif
