--- processinfo/processinfo_loopstep.h
@@ -1,6 +1,0 @@
-#ifndef _PROCESSINFO_LOOPSTEP_H
-#define _PROCESSINFO_LOOPSTEP_H
-
-int processinfo_loopstep(PROCESSINFO *processinfo);
-
-#endif
