--- processinfo/processinfo_procdirname.h
@@ -1,6 +1,0 @@
-#ifndef _PROCESSINFO_PROCDIRNAME_H
-#define _PROCESSINFO_PROCDIRNAME_H
-
-errno_t processinfo_procdirname(char *procdname);
-
-#endif
