--- processinfo/processinfo_shm_close.h
@@ -1,6 +1,0 @@
-#ifndef _PROCESSINFO_SHM_CLOSE_H
-#define _PROCESSINFO_SHM_CLOSE_H
-
-int processinfo_shm_close(PROCESSINFO *pinfo, int fd);
-
-#endif
