--- processinfo/processinfo_shm_create.h
@@ -1,6 +1,0 @@
-#ifndef _PROCESSINFO_SHM_CREATE_H
-#define _PROCESSINFO_SHM_CREATE_H
-
-PROCESSINFO *processinfo_shm_create(const char *pname, int CTRLval);
-
-#endif
