--- processinfo/processinfo_shm_link.h
@@ -1,6 +1,0 @@
-#ifndef _PROCESSINFO_SHM_LINK_H
-#define _PROCESSINFO_SHM_LINK_H
-
-PROCESSINFO *processinfo_shm_link(const char *pname, int *fd);
-
-#endif
