--- processinfo/processinfo_shm_list_create.h
@@ -1,6 +1,0 @@
-#ifndef _PROCESSINFO_SHM_LIST_CREATE_H
-#define _PROCESSINFO_SHM_LIST_CREATE_H
-
-long processinfo_shm_list_create();
-
-#endif
